Change ingredient amount column type to decimal
To better track things like 1/4 cups, etc.

+++ database/migrations/2024_10_31_001758_create_ingredient_recipe_pivot_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
         // duplication in the ingredients table, and this is a solution to that problem.
         Schema::create('ingredient_recipe', function (Blueprint $table) {
             // The amount of this ingredient that should go into the recipe
-            $table->unsignedSmallInteger('amount');
+            $table->decimal('amount');
             // The units for the amount to put into the recipe. E.g. cups, grams, ounces, etc. Allow nullable to
             // indicate when a unit doesn't make sense (e.g. "3 potatoes" doesn't need a unit).
             $table->string('unit')->nullable();
